+++ app/src/main/java/com/example/househomey/MainActivity.java
@@ -42,7 +42,7 @@ public class MainActivity extends AppCompatActivity {
         }
 
         // Init home fragment
-        navigateToFragmentPage(this, new HomeBaseStateFragment());
+        navigateToFragmentPage(this, new HomeFragment());
 
         BottomNavigationView bottomNavigationView = findViewById(R.id.bottom_navigation);
         bottomNavigationView.setOnItemSelectedListener(item -> {
@@ -50,13 +50,13 @@ public class MainActivity extends AppCompatActivity {
             Fragment fragment;
             if (id == R.id.action_home) {
                 // Go to Home page
-                fragment = new HomeBaseStateFragment();
+                fragment = new HomeFragment();
             } else if (id == R.id.action_add) {
                 // Go to Add Item page
                 fragment = new AddItemFragment();
             } else {
                 // TODO: Go to Profile Page
-                fragment = new HomeBaseStateFragment();
+                fragment = new HomeFragment();
             }
             navigateToFragmentPage(this, fragment);
             return true;
