+++ app/src/main/java/com/example/househomey/filter/model/FilterCallback.java
@@ -3,7 +3,7 @@ package com.example.househomey.filter.model;
 
 /**
  * Callback interface for handling filter changes between fragments.
- * @see com.example.househomey.HomeBaseStateFragment
+ * @see com.example.househomey.HomeFragment
  */
 public interface FilterCallback {
     void onFilterApplied(Filter filter);
